BigInteger issue fixed

+++ app/db/app_models.py
@@ -1,5 +1,5 @@
 # app/db/app_models.py
-from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean, BigInt, create_engine
+from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean, BigInteger, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from datetime import datetime
@@ -40,8 +40,8 @@ class ChatLog(Base):
     
     # --- NEW METRICS (Measurable Proof) ---
     resolved_latest_ds = Column(String, nullable=True)
-    row_count = Column(BigInt, nullable=True)
-    execution_ms = Column(BigInt, nullable=True)
+    row_count = Column(BigInteger, nullable=True)
+    execution_ms = Column(BigInteger, nullable=True)
     visual_type = Column(String, nullable=True)
     correction_attempts = Column(Integer, default=0)
     tables_used = Column(Text, nullable=True)
